requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
bcrypt==4.1.2
python-dotenv==1.0.0
itsdangerous==2.2.0
